+++ src/utils/core.jsx
@@ -475,9 +475,39 @@ export const usePageFocused = () => {
 			window.removeEventListener('blur', handleFocus)
 			window.removeEventListener('focus', handleFocus)
 		}
-	}, [])
+	}, []);
 
-	return isFocused
+	return isFocused;
+};
+
+/**!
+ *
+ * `useSignalsPageFocused()` ReactJS hook
+ */
+
+export const useSignalsPageFocused = () => {
+	const [isFocused, setIsFocused] = useSignalsState(() => {
+		if (typeof window !== 'undefined') {
+			return document.hasFocus()
+		}
+		return false
+	})
+
+	const handleFocus = () => {
+		setIsFocused(document.hasFocus())
+	}
+
+	useSignalsEffect(() => {
+		window.addEventListener('blur', handleFocus)
+		window.addEventListener('focus', handleFocus)
+
+		return () => {
+			window.removeEventListener('blur', handleFocus)
+			window.removeEventListener('focus', handleFocus)
+		}
+	}, []);
+
+	return isFocused;
 }
 
 /**!
@@ -486,9 +516,34 @@ export const usePageFocused = () => {
 
 export const useBeforePageUnload = (
 	callback = () => undefined,
-	{ when, message }
+	{ when = false, message = "" }
 ) => {
 	useEffect(() => {
+		const handleBeforeUnload = (event) => {
+			event.preventDefault()
+			callback.call(null, event.target)
+			event.returnValue = message
+			return message
+		}
+
+		if (when) {
+			window.addEventListener('beforeunload', handleBeforeUnload)
+		}
+
+		return () => window.removeEventListener('beforeunload', handleBeforeUnload)
+		/* eslint-disable-next-line react-hooks/exhaustive-deps */
+	}, [when, message])
+};
+
+/**!
+ * `useSignalsBeforePageUnload()` ReactJS hook
+ */
+
+ export const useSignalsBeforePageUnload = (
+	callback = () => undefined,
+	{ when = false, message = "" }
+) => {
+	useSignalsEffect(() => {
 		const handleBeforeUnload = (event) => {
 			event.preventDefault()
 			callback.call(null, event.target)
@@ -601,6 +656,30 @@ export const useIsDOMElementVisibleOnScreen = (options = { root: null, threshold
 	const [isIntersecting, setIsIntersecting] = useState(false)
 
 	useEffect(() => {
+		const domElement = domElementRef.current;
+		const iterator = (entry) => {
+			return setIsIntersecting(() => entry.isIntersecting)
+		}
+		const callback = (entries) => entries.forEach(iterator)
+		const observer = new window.IntersectionObserver(callback, options)
+
+		if (domElement) {
+			observer.observe(domElement)
+		}
+		return () => domElement && observer.unobserve(domElement)
+	}, [options])
+
+	return [isIntersecting, domElementRef];
+};
+
+/**!
+ * `useSignalsIsDOMElementVisibleOnScreen()` ReactJS hook
+ */
+ export const useSignalsIsDOMElementVisibleOnScreen = (options = { root: null, threshold: 0 }) => {
+	const domElementRef = useRef(null);
+	const [isIntersecting, setIsIntersecting] = useSignalsState(false)
+
+	useSignalsEffect(() => {
 		const domElement = domElementRef.current;
 		const iterator = (entry) => {
 			return setIsIntersecting(() => entry.isIntersecting)